refactor: 상세 화면 컴포넌트 분리 및 파일 구조 최적화
- `AfternoteDetailComponents.kt` 파일을 삭제하고 역할을 기준으로 개별 파일로 분리
- `DetailSection`, `DetailSectionHeader`, `DetailCard`를 `DetailSectionLayout.kt`로 이동하여 레이아웃 구조 캡슐화
- `DetailInfoRow`를 별도 파일로 분리하여 단일 행 컴포넌트로서의 재사용성 강화
- 도메인 전용 섹션인 `MessageSection`과 `ProcessingMethodsSection`을 각각 독립된 파일로 분리
- 각 파일별로 Preview 코드 및 필요한 리소스 임포트 정리 (예: `com.afternote.core.ui.R` 별칭 사용 등)
